pyproject.toml
[project]
name = "follow-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.43.1",
    "lxml>=6.0.0",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "python-fasthtml>=0.12.24",
    "requests>=2.32.5",
]
